pyproject.toml
[tool.poetry]
name = "emotional-friend-bot"
version = "0.1.0"
description = "Telegram bot for emotional support"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.13.1"
openai = "1.51.0"
apscheduler = "3.11.0"
python-dotenv = "1.0.1"
aiosqlite = "0.20.0"
tzlocal = "5.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"